code structure reorganization
